requirements.txt
streamlit==0.64.0
selenium==3.141.0
pandas==1.0.5
numpy==1.18.5
Pillow==7.2.0
